Split BoardCanvasView and group BoardCanvas/ by concern
Extract nested types out of BoardCanvasView.swift into their own files
under new Elements/ and Import/ subfolders. BoardCanvasView.swift
shrinks from ~2470 lines to ~2120, containing only the main view.

Extracted into Features/BoardCanvas/Elements/:
- PlacedImage, PlacedText (model structs)
- TextElementView, FileImageView (rendering views)
- ImageCache (UIImage NSCache)

Extracted into Features/BoardCanvas/Import/:
- CanvasDropDelegate (drag-drop adoption)
- ItemProviderHelpers (loadURLsFromProviders + NSItemProvider extension)
- InsertFileControl moved in alongside

Also fold the toolbar into BoardCanvas/Tools/:
- CanvasTool, CanvasToolBehavior, CanvasStatusBar move into Tools/
- Features/HUD/CanvasToolbar.swift moves to Tools/CanvasToolbar.swift
- Drop the now-empty Features/HUD/ folder

Update architecture-frontend.md and CLAUDE.md to reflect the new tree.
