+++ CLAUDE.md
@@ -52,11 +52,13 @@ SuperCoolArtReferenceTool/
 ├── DesignSystem/               Colors, shared design tokens
 ├── Features/
 │   ├── BoardCanvas/            Infinite canvas (main view + supporting modules)
+│   │   ├── Elements/           Placed image/text models + their rendering views, image cache
 │   │   ├── Gestures/           Pinch / two-finger pan installers
+│   │   ├── Import/             Drag/drop delegate, item-provider helpers, insert-file button
 │   │   ├── Selection/          Selection state, marquee, handles, action bar
-│   │   └── Settings/           Settings sheet + button
-│   ├── FilePicker/             Landing screen, recent-board management
-│   └── HUD/                    Persistent toolbar
+│   │   ├── Settings/           Settings sheet + button
+│   │   └── Tools/              Toolbar, tool definitions, tool behavior, status bar
+│   └── FilePicker/             Landing screen, recent-board management
 └── Persistence/                In-memory spatial store, canvas models, services
 ```
 
@@ -70,9 +72,9 @@ SuperCoolArtReferenceTool/
 | `Persistence/LocalCanvasService.swift` | Service layer wrapping LocalBoardStore |
 | `App/BoardArchiver.swift` | Import/export `.refboard` packages (JSON manifest + assets/) |
 | `App/ContentView.swift` | Main canvas container, toolbar layout, export/import UI |
-| `Features/HUD/CanvasToolbar.swift` | Left/right toolbar with tool selection, undo/redo |
+| `Features/BoardCanvas/Tools/CanvasToolbar.swift` | Left/right toolbar with tool selection, undo/redo |
 | `Features/FilePicker/FilePickerView.swift` | Landing screen with drag-and-drop/browse for initial file import |
-| `Features/BoardCanvas/InsertFileControl.swift` | Reusable file-import button used by the canvas |
+| `Features/BoardCanvas/Import/InsertFileControl.swift` | Reusable file-import button used by the canvas |
 | `DesignSystem/Colors.swift` | Color palette: primary (#191919), secondary (#535353), tertiary (#86B8FE), text (#FFFFFF) |
 
 ### Coordinate Systems
@@ -101,5 +103,5 @@ Document only finalized implementations, not speculative architecture.
 
 ## Known Issues
 
-- File loading logic duplicated between `Features/BoardCanvas/BoardCanvasView.swift` and `Features/BoardCanvas/InsertFileControl.swift`.
+- File loading logic duplicated between `Features/BoardCanvas/BoardCanvasView.swift` and `Features/BoardCanvas/Import/InsertFileControl.swift`.
 - Pointer and group tools in toolbar are not yet connected to canvas behavior.
