+++ SuperCoolArtReferenceTool/Features/BoardCanvas/Elements/PlacedImage.swift
@@ -1,0 +1,8 @@
+import SwiftUI
+
+struct PlacedImage: Identifiable {
+    let id: UUID
+    let url: URL
+    var worldRect: CGRect
+    var zIndex: Int
+}
