+++ SuperCoolArtReferenceTool/Features/BoardCanvas/Import/ItemProviderHelpers.swift
@@ -1,0 +1,65 @@
+import Foundation
+import UniformTypeIdentifiers
+
+func loadURLsFromProviders(_ providers: [NSItemProvider], preferredTypes: [UTType]) async -> [URL] {
+    await withTaskGroup(of: URL?.self) { group in
+        for provider in providers {
+            if let firstType = preferredTypes.first(where: { provider.hasItemConformingToTypeIdentifier($0.identifier) }) {
+                group.addTask {
+                    if let url = await provider.loadFileURLCompat(for: firstType) { return url }
+                    if let dataURL = await provider.loadDataAsTempFileCompat(for: firstType) { return dataURL }
+                    return nil
+                }
+            }
+        }
+        var results: [URL] = []
+        for await url in group { if let url { results.append(url) } }
+        return results
+    }
+}
+
+extension NSItemProvider {
+    func loadFileURLCompat(for type: UTType) async -> URL? {
+        await withCheckedContinuation { continuation in
+            self.loadFileRepresentation(forTypeIdentifier: type.identifier) { url, _ in
+                guard let url else {
+                    continuation.resume(returning: nil)
+                    return
+                }
+                // The provided URL is temporary and deleted when this callback returns,
+                // so copy it to a stable temp location before resuming.
+                let ext = url.pathExtension.isEmpty ? (type.preferredFilenameExtension ?? "dat") : url.pathExtension
+                let dest = FileManager.default.temporaryDirectory
+                    .appendingPathComponent(UUID().uuidString)
+                    .appendingPathExtension(ext)
+                do {
+                    try FileManager.default.copyItem(at: url, to: dest)
+                    continuation.resume(returning: dest)
+                } catch {
+                    continuation.resume(returning: nil)
+                }
+            }
+        }
+    }
+
+    func loadDataAsTempFileCompat(for type: UTType) async -> URL? {
+        await withCheckedContinuation { continuation in
+            self.loadDataRepresentation(forTypeIdentifier: type.identifier) { data, _ in
+                guard let data else {
+                    continuation.resume(returning: nil)
+                    return
+                }
+                let tempDir = URL(fileURLWithPath: NSTemporaryDirectory(), isDirectory: true)
+                let ext = type.preferredFilenameExtension ?? "dat"
+                let filename = UUID().uuidString + "." + ext
+                let url = tempDir.appendingPathComponent(filename)
+                do {
+                    try data.write(to: url, options: [.atomic])
+                    continuation.resume(returning: url)
+                } catch {
+                    continuation.resume(returning: nil)
+                }
+            }
+        }
+    }
+}
